# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yapaia Go Core API",
    "version": "0.0.1",
    "description": "Automatisch generiert aus den tatsächlich registrierten Fastify-Routen (apps/core/src/index.ts, onRouteHook) und den @yapaia/shared-JSON-Schemas (packages/shared/src/schemas/). Quelle der Wahrheit ist der Code, nicht dieses Dokument -- bei Verdacht auf Abweichung: `pnpm openapi:check` (CI-Gate \"Spec aktuell\") bzw. docs/03-api-spec.md für den erzählenden Überblick. WebSocket (/ws/v1) und MQTT sind hier NICHT modelliert (OpenAPI eignet sich dafür nicht) -- siehe docs/03-api-spec.md §3/§4."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Core-API (relative Pfade -- funktioniert unverändert hinter HA-Ingress, W-15)"
    }
  ],
...
  "paths": {
...
    "/api/v1/position": {
      "get": {
        "summary": "Letzte bekannte Position",
        "tags": [
          "Position"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Position"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/position/sources": {
      "get": {
        "summary": "Verfügbare Positionsquellen + aktive Quelle",
        "tags": [
          "Position"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/position/source": {
      "put": {
        "summary": "Positionsquelle erzwingen (gpsd|browser|auto|simulator)",
        "tags": [
          "Position"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/position/browser": {
      "post": {
        "summary": "Browser-Geolocation-Fix melden",
        "tags": [
          "Position"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Position"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/map/regions": {
      "post": {
        "summary": "Region-Download starten (202 + job_id)",
        "tags": [
          "Karten"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Installierte Kartenregionen",
        "tags": [
          "Karten"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/tiles/{regionParam}": {
      "get": {
        "summary": "PMTiles-Kachel (HTTP-Range-Support)",
        "tags": [
          "Karten"
        ],
        "parameters": [
          {
            "name": "regionParam",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binärdaten (kein JSON-Schema veröffentlicht)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/map/build-status": {
      "get": {
        "summary": "GET /api/v1/map/build-status",
        "tags": [
          "Sonstige"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/map/styles": {
      "get": {
        "summary": "Verfügbare MapLibre-Styles",
        "tags": [
          "Karten"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/map/styles/{id}": {
      "get": {
        "summary": "MapLibre-Style-JSON (lokale Tile-URLs)",
        "tags": [
          "Karten"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort (Schema noch nicht kuratiert -- siehe apps/core/src/openapi/paths.ts)"
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Position": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "description": "Latitude in degrees (-90 to 90)"
          },
          "lon": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "description": "Longitude in degrees (-180 to 180)"
          },
          "alt": {
            "type": [
              "number",
              "null"
            ],
            "description": "Altitude in meters above mean sea level"
          },
          "speed": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "description": "Speed in m/s over ground"
          },
          "heading": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "maximum": 360,
            "description": "Heading in degrees (0 = North)"
          },
          "accuracy": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "description": "Horizontal accuracy in meters (HDOP-based)"
          },
          "source": {
            "type": "string",
            "enum": [
              "gpsd",
              "browser",
              "simulator",
              "ha_tracker"
            ],
            "description": "Source of position data"
          },
          "fix": {
            "type": "string",
            "enum": [
              "none",
              "2d",
              "3d"
            ],
            "description": "GPS fix type"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp in ISO 8601 UTC"
          }
        },
        "required": [
          "lat",
          "lon",
          "alt",
          "speed",
          "heading",
          "accuracy",
          "source",
          "fix",
          "ts"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code identifier"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": [
                  "object",
                  "null"
                ],
                "description": "Optional additional error context"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false,
            "description": "Error object with code, message, and optional details"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}